Ennemi se déplace maintenant en 2i en fonction de la grid

+++ Project1/EnemyManager.cpp
@@ -26,10 +26,11 @@ void EnemyManager::draw(RenderWindow& window)
 	}
 }
 
-void EnemyManager::createMGSPatrol(float posX, float posY, Vector2f p1, Vector2f p2, Vector2f p3)
+void EnemyManager::createMGSPatrol(float posX, float posY, Vector2i p1, Vector2i p2, Vector2i p3)
 {
-	m_mgs_enemies.push_back(make_unique<PatrolMGS>(posX, posY,p1,p2,p3));
+	m_mgs_enemies.push_back(make_unique<PatrolMGS>(posX, posY, p1, p2, p3));
 }
+
 
 void EnemyManager::setMenacedState()
 {
